requirements.txt
pyaml
boto
boto3
awscli
pywinrm
setuptools-rust
rust
cryptography
ansible

